+++ scripts/register-settings.mjs
@@ -6,6 +6,7 @@ import { log } from "./lib.mjs";
 import { EnableMenu } from "./apps/enable-menu.mjs";
 import { ExportDialog, ImportDialog } from "./export.mjs";
 import { PlayerHandbookMenu } from "./apps/player-handbook.mjs";
+import { getOverrideCompendiumOptions } from "./presentation/override-compendium.mjs";
 
 
 export function initVersionSetting() {
@@ -159,6 +160,7 @@ export function initSettings() {
         }
       );
 
+
       if (game.modules.get("quick-insert")) {
         game.settings.register(MODULE_NAME, SETTINGS.filterQuickInsert, 
             {
@@ -225,4 +227,22 @@ export function initSettings() {
     ) 
 
     console.groupEnd()
+}
+
+export function registerCompendiumOverrideSetting() {
+    //Handles settings which are only registered once Foundry is ready
+
+    game.settings.register(MODULE_NAME, SETTINGS.overrideCompendium,
+        {
+            name: "Handbook Override Compendium",
+            hint: "Compendium to search for custom Class, Subclass, Species, and Background Journal Pages. If the desired compendium isn't present, reload Foundry",
+            config: true,
+            scope: "world",
+            default: "none",
+            type: String, 
+            restricted: true,
+            requiresReload: false,
+            choices: getOverrideCompendiumOptions()
+        }
+    )
 }