+++ scripts/settings.mjs
@@ -191,7 +191,8 @@ export const SETTINGS = {
     filterHeromancer: "filterHeromancer",
     injectCompendiumButtons: "injectCompendiumButtons",
     injectItemButton: "injectItemButton",
-    playerHandbookOptions: "playerHandbookOptions"
+    playerHandbookOptions: "playerHandbookOptions",
+    overrideCompendium: "overrideCompendium"
 }
 
 /**